+++ lustre/ldlm/ldlm_request.c
@@ -1801,10 +1801,10 @@ int ldlm_cancel_lru(struct ldlm_namespace *ns, int nr,
  * list.
  */
 int ldlm_cancel_resource_local(struct ldlm_resource *res,
-                               cfs_list_t *cancels,
-                               ldlm_policy_data_t *policy,
-                               ldlm_mode_t mode, int lock_flags,
-                               ldlm_cancel_flags_t cancel_flags, void *opaque)
+			       cfs_list_t *cancels,
+			       ldlm_policy_data_t *policy,
+			       ldlm_mode_t mode, __u64 lock_flags,
+			       ldlm_cancel_flags_t cancel_flags, void *opaque)
 {
         struct ldlm_lock *lock;
         int count = 0;
